requirements.txt
bcrypt
cryptography
python-jose
python-multipart
openai
SQLAlchemy
fastapi
psycopg2-binary
passlib
uvicorn
python-dotenv
python-dateutil
pydantic
pydantic[email]
langchain>=0.2.0
langchain-openai
langchain-community>=0.0.17
faiss-cpu
pypdf
python-docx
pandas
aiohttp>=3.8.3,<4.0.0

